requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asyncio-dgram==2.2.0
attrs==25.4.0
audioop-lts==0.2.2
discord.py==2.6.4
dnspython==2.8.0
frozenlist==1.8.0
idna==3.11
mcstatus==12.0.6
multidict==6.7.0
propcache==0.4.1
python-dotenv==1.2.1
setuptools==80.9.0
wakeonlan==3.1.0
yarl==1.22.0
